refactor(core): migrate OCP.AppConfig from jQuery to axios
